사용 옵션 수정
