--- app/types/base.ts
@@ -1,8 +1,0 @@
-export interface AuthError {
-  error: string;
-  message: string;
-}
-
-export type BaseServerResposne<T, E = AuthError> =
-  | { isSuccess: true; data: T }
-  | { isSuccess: false; error: E };
